Change from using the unofficial winapi bindings, to using the official windows bindings from Microsoft.
I have also a PR open at winit: https://github.com/rust-windowing/winit/pull/2057
They suggested that parking_lot (dependency of winit) should also be switched over to windows-rs.
For this port I am using the more lightweight and lower level windows-sys (which is closer to winapi).
For me build times are faster with windows-rs 🎉 
Before(winapi 1.19s, total time 2.2s):
![winapi_parking_lot](https://user-images.githubusercontent.com/26170289/148648990-cb504f50-63c6-41d5-87f8-b110186b3b59.png)
After(windows-sys 0.94s, total time 1.7s):
![windows_parking_lot](https://user-images.githubusercontent.com/26170289/148648996-6eab574b-fe82-4d18-b645-e939312fafe5.png)