@bjorn3 2022-01-08
Several other crates also compile faster in the second compile. Are you sure this isn't the cpu changing frequency?