@clemenswasser 2022-01-08
> Several other crates also compile faster in the second compile. Are you sure this isn't the cpu changing frequency?

It is deterministically faster.
When I build it multiple times, it's always winit ~2s and windows-rs ~1.7s.