@madsmtm 2022-01-13
Note that this might be an MSRV bump (at least when the `windows` crate had an MSRV, see https://github.com/microsoft/windows-rs/issues/1426#issuecomment-1011795357)